Follow up 9b4b01061acdcb027d89d9e505f00278c3c82f51
Restructured confs further

+++ src/Shared/Util/Log.cs
@@ -9,7 +9,7 @@ namespace Aura.Shared.Util
 	/// <summary>
 	/// All log levels. Also used as bitmask, for hiding.
 	/// </summary>
-	public enum LogLevel : short
+	public enum LogLevel : int
 	{
 		Info = 0x0001,
 		Warning = 0x0002,
